bot_service/requirements/requirements.txt
# Telegram Bot
aiogram==2.25.1

# Database
sqlalchemy==2.0.41
asyncpg==0.29.0
alembic==1.13.1

# HTTP client
aiohttp==3.8.6

# Environment and configuration
python-dotenv==1.0.1

# QR Code generation
qrcode==7.4.2
pypng==0.20220715.0

# Utilities
pydantic==2.11.7
pydantic_core==2.33.2
